requirements.txt
Pillow 
OpenCV